OneDrive/Desktop/New folder (2)/requirements.txt
absl-py==1.0.0
arabic-reshaper==2.1.3
asgiref==3.4.1
astor==0.8.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
dj-database-url==0.5.0
Django==3.2.10
django-ckeditor==6.3.2
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.0.0
docopt==0.6.2
future==0.18.2
gast==0.5.3
google-pasta==0.2.0
grpcio==1.45.0
gunicorn==20.1.0
h5py==2.10.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
imutils==0.5.4
joblib==1.1.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.1.1
mysql-connector-python==8.0.27
num2words==0.5.10
numpy==1.19.2
opencv-python==4.1.1.26
pandas==0.25.3
Pillow==8.4.0
prettytable==3.3.0
protobuf==3.19.3
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pywinpty==2.0.5
PyYAML==6.0
reportlab==3.6.6
requests==2.27.1
scikit-learn==0.22.2.post1
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.13.3
tornado==6.1
traitlets==5.2.0
twilio==7.4.0
typing-extensions==4.0.1
umpy==0.1.0
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
whitenoise==6.2.0
wrapt==1.14.0
xhtml2pdf==0.2.5
zipp==3.8.0
